refactor: mv item record parsing to item page
